requirements.txt
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.14.0
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.25.0
numpy>=1.21.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.7